Changed signature of the Operation call operator
It is now possible to return a different buffer from an operation.
This should make it easier to prepare data (e.g. between frames) and
then just pass it back to the rendering sequences.

This is also required for alpha blending where each Operation will
receive it's own render buffer.